+++ src/command/command.service.ts
@@ -71,6 +71,7 @@ export class CommandService {
   setInitialCommandState() {
     this.commandState = {
       arena: '',
+      shoutoutUsers: CONFIG.get().autoShoutouts,
       boughtSquares: {},
       first: '',
       challengeQueue: [],